moved session_path from config to session_file util
