+++ tests/cli/test_session.py
@@ -16,14 +16,18 @@ def mock_specified_session_name():
     with patch.object(PromptSession, "prompt", return_value=SPECIFIED_SESSION_NAME) as specified_session_name:
         yield specified_session_name
 
+@pytest.fixture
+def mock_sessions_path(tmp_path):
+    with patch("goose.config.SESSIONS_PATH", tmp_path) as mock_path:
+        yield mock_path
 
 @pytest.fixture
-def create_session_with_mock_configs(mock_sessions_path, exchange_factory, profile_factory):
+def create_session_with_mock_configs(exchange_factory, profile_factory, tmp_path):
     with patch("goose.cli.session.build_exchange", return_value=exchange_factory()), patch(
         "goose.cli.session.load_profile", return_value=profile_factory()
     ), patch("goose.cli.session.SessionNotifier") as mock_session_notifier, patch(
         "goose.cli.session.load_provider", return_value="provider"
-    ):
+    ), patch("goose.utils.session_file.SESSIONS_PATH", tmp_path):
         mock_session_notifier.return_value = MagicMock()
 
         def create_session(session_attributes: dict = {}):
